Template: restructure logic in `econozel_theme_compat_reset_post`
In order to know in a template file whether we are in theme-compat mode,
the call to `econozel_set_theme_compat_active()` has to happen before
calling `econozel_buffer_template_part()`. To do is, I moved the call to
buffer the template part within the dummy post constructor, which now
takes an array of template part arguments as an additional parameter.

Now the theme-compat mode is defined before the template file is loaded.