Map meta caps for the post_tag taxonomy
This requires rerouting taxonomy caps from the default caps, which are
'edit_posts' for assigning terms and 'manage_categories' for all other
caps. Mapping these particular caps is off course a no-go since they are
used accross WP for a multitude of cap-checks. Therefore, the caps for the
post_tag taxonomy are now uniquely identifiable, while we defer to the
original caps when the user is not an Econozel Editor.

See #3