Enable listing recent or archived Articles
The new 'econozel_archive' query parameter enables to distinguish posts by
publication. Recent posts are considered those that are published in the
latest Edition or those not published in any Edition. Archived posts are
those published in any other Edition, which are considered the archive.

Query options are provided in the Econozel Recent Articles widget, renamed
to Econozel Articles. An additional instance of that widget is added to
the Econozel Home page displaying the recently archived posts in contrast
to the recent published posts displayed in the list of Latest Articles.