Introduce Edition cover photos
Enables selecting a media attachment as an Edition's featured image. For
now this is implemented as an Edition's Article's post thumbnail when the
Article has no featured image of its own.

Using the image as the actual Edition page's post thumbnail requires some
more time to figure out how to implement.